requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-dotenv
passlib[bcrypt]
requests
